pyproject.toml
[project]
name = "buckshot-lambda"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
	"speciesnet>=5.0.3",
	"opencv-python-headless",
	"setuptools<81.0.0",
	"python-dotenv",
	"boto3",
	"boto3-stubs[lambda]",
	"boto3-stubs[s3]",
]

[tool.uv]
exclude-dependencies = ["opencv-python"]
